Create Question11.py
his Python script attempts to find the second largest number in a list by subtracting 1 from the maximum value of the list. However, this method does not reliably find the second largest element because the second largest number may not be exactly one less than the maximum.

Note: To accurately find the second largest number, more robust methods should be used, such as removing the maximum and then finding the next largest number.

🧪 Code Provided
lst = [1, 26, 4, 20, 23, 32, 33]
a = max(lst) - 1
print("Second largest number in the list is:", a)

⚠️ Important

This code will not always return the correct second largest number.

For example, if the list is [1, 2, 3, 5], subtracting 1 from the max (5) gives 4, which is not in the list.

To correctly find the second largest, consider alternative approaches like:

Removing the maximum and taking the max of the remainder.

Sorting the list and selecting the second last element.